Fix LD_PRELOAD handling
Previous versions really *pre*loaded libraries, before the binaries.
However, this is not only different to Glibc (and might result in
changed behaviour in case of symbol overloading), it also can break
applications with TLS: preloaded libs or its dependencies would first
allocate TLS (next to %fs), but gcc already expected this to be reserved
for the main binary.