Option to stop process during update
According to the Intel manual (cross processor code modification) it is
required to stop the thread when modifing its code.
While all tests suggests that the type of changes done by luci don't
require this limitation, this patch adds an option to stop (SIGSTOP) the process
during rewriting the code and continues (SIGCONT) it.
However, this might put a foreground process into the background during
the update, so in Bash you should better disable job control when using
this feature.